requirements.txt
torch
gradio
google-generativeai
python-dotenv